Internal parameters + simple execution parameter
Internal parameters are now used to automatically keep track of certain
information useful for verification or test execution. These can be
overriden if the user supplies the parameters.

Also added simple stub calls for customizing wether checksum should be
used for verification and to specify a state-number to be able to, in
the future, verify a test case against different result sets